containerinit: set default route on eth0 only
As we are rendering /etc/network/interfaces we add rules that replace
the default route per iface stanza to render but this has the
unfortunate effect of moving the default route from eth0..eth1..ethN.
This breaks the previous Juju 1.25.x assumptions/model where the default
route was always on eth0 in the container and juju-br0 on the host. This
change ensures that the default gateway stays on eth0. But this change
is mostly mitigation against changing the previous 1.x Juju behaviour.
The correct change is to discover what the default route is on the host
and propagate that to the correct interface in the container.

Fixes [LP:#1576674](https://bugs.launchpad.net/juju-core/+bug/1576674)