containerinit: don't assume eth0 is the gateway
Remove the assumption that eth0 must be the default gateway. When
rendering /etc/network/interfaces validate whether any default gateway
information that is present is appropriate for the iface stanza that is
currently being rendered.

Only one 'gateway <addr>' can be present in /etc/network/interfaces,
additional entries must have 'metric' options to disambiguate them. As
we don't want to make up arbitrary metrics we write an entry for the
first case of:

  ifaceNetworkAddress.Contains(gatewayIPAddress)

or where we see an interface configured as DHCP and we assume we'll get
a default route with the lease.

Fixes [LP:#1602054](https://bugs.launchpad.net/juju-core/+bug/1602054)